refactor : 폴더구조 변경
SL-184
