+++ backend/src/main/java/com/devdive/backend/hello/adaptor/in/web/HelloController.java
@@ -1,7 +1,7 @@
 package com.devdive.backend.hello.adaptor.in.web;
 
-import com.devdive.backend.security.core.AuthenticationPrincipal;
-import com.devdive.backend.security.core.UserDetails;
+import com.devdive.backend.security.core.resolver.AuthenticationPrincipal;
+import com.devdive.backend.security.authentication.domain.UserDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
