+++ backend/src/main/java/com/devdive/backend/security/config/FilterConfig.java
@@ -3,7 +3,7 @@ package com.devdive.backend.security.config;
 import com.devdive.backend.auth.application.service.jwt.JwtProvider;
 import com.devdive.backend.security.config.builder.HttpSecurity;
 import com.devdive.backend.security.filter.FilterChainProxy;
-import com.devdive.backend.security.filter.SecurityFilterChain;
+import com.devdive.backend.security.core.SecurityFilterChain;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
