--- backend/src/main/java/com/devdive/backend/security/core/UserDetailsService.java
@@ -1,6 +1,0 @@
-package com.devdive.backend.security.core;
-
-public interface UserDetailsService<I> {
-
-    UserDetails findMemberByEmail(I mail);
-}
