+++ backend/src/main/java/com/devdive/backend/security/core/resolver/SecurityContextResolver.java
@@ -1,6 +1,9 @@
-package com.devdive.backend.security.core;
+package com.devdive.backend.security.core.resolver;
 
-import com.devdive.backend.security.authentication.AnonymousAuthenticationToken;
+import com.devdive.backend.security.authentication.Authentication;
+import com.devdive.backend.security.authentication.domain.UserDetails;
+import com.devdive.backend.security.authentication.token.AnonymousAuthenticationToken;
+import com.devdive.backend.security.core.securitycontext.SecurityContextHolder;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
