+++ backend/src/main/java/com/devdive/backend/security/core/resolver/config/ResolverConfig.java
@@ -1,5 +1,6 @@
-package com.devdive.backend.security.core;
+package com.devdive.backend.security.core.resolver.config;
 
+import com.devdive.backend.security.core.resolver.SecurityContextResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
