+++ backend/src/main/java/com/devdive/backend/security/core/securitycontext/SecurityContext.java
@@ -1,4 +1,6 @@
-package com.devdive.backend.security.core;
+package com.devdive.backend.security.core.securitycontext;
+
+import com.devdive.backend.security.authentication.Authentication;
 
 public interface SecurityContext {
     Authentication getAuthentication();
