+++ backend/src/main/java/com/devdive/backend/security/filter/AnonymousAuthenticationFilter.java
@@ -1,8 +1,8 @@
 package com.devdive.backend.security.filter;
 
-import com.devdive.backend.security.authentication.AnonymousAuthenticationToken;
-import com.devdive.backend.security.core.Authentication;
-import com.devdive.backend.security.core.SecurityContextHolder;
+import com.devdive.backend.security.authentication.token.AnonymousAuthenticationToken;
+import com.devdive.backend.security.authentication.Authentication;
+import com.devdive.backend.security.core.securitycontext.SecurityContextHolder;
 import jakarta.servlet.*;
 
 import java.io.IOException;
