+++ backend/src/test/java/com/devdive/backend/security/access/manager/AccessDecisionManagerImpTest.java
@@ -1,7 +1,7 @@
 package com.devdive.backend.security.access.manager;
 
 import com.devdive.backend.security.access.AccessDecisionVoter;
-import com.devdive.backend.security.core.Authentication;
+import com.devdive.backend.security.authentication.Authentication;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
