+++ backend/src/test/java/com/devdive/backend/security/access/voter/RoleVoterTest.java
@@ -1,6 +1,6 @@
 package com.devdive.backend.security.access.voter;
 
-import com.devdive.backend.security.core.Authentication;
+import com.devdive.backend.security.authentication.Authentication;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
