+++ backend/src/test/java/com/devdive/backend/security/authentication/AnonymousAuthenticationTokenTest.java
@@ -1,5 +1,6 @@
 package com.devdive.backend.security.authentication;
 
+import com.devdive.backend.security.authentication.token.AnonymousAuthenticationToken;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
