+++ backend/src/test/java/com/devdive/backend/security/core/DefaultUserDetailsServiceTest.java
@@ -1,5 +1,9 @@
 package com.devdive.backend.security.core;
 
+import com.devdive.backend.security.authentication.application.port.out.LoadMemberPort;
+import com.devdive.backend.security.authentication.application.service.DefaultUserDetailsService;
+import com.devdive.backend.security.authentication.domain.UserDetails;
+import com.devdive.backend.security.authentication.application.port.in.UserDetailsService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
