+++ backend/src/test/java/com/devdive/backend/security/core/SecurityContextHolderStrategyImpTest.java
@@ -1,5 +1,8 @@
 package com.devdive.backend.security.core;
 
+import com.devdive.backend.security.core.securitycontext.SecurityContextHolderStrategy;
+import com.devdive.backend.security.core.securitycontext.persistent.SecurityContextHolderStrategyImp;
+import com.devdive.backend.security.core.securitycontext.SecurityContextImp;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.*;
