+++ backend/src/test/java/com/devdive/backend/security/filter/TokenAuthenticationFilterTest.java
@@ -2,8 +2,8 @@ package com.devdive.backend.security.filter;
 
 import com.devdive.backend.auth.application.service.jwt.JwtProvider;
 import com.devdive.backend.security.core.AuthenticationCache;
-import com.devdive.backend.security.core.UserDetails;
-import com.devdive.backend.security.core.UserDetailsService;
+import com.devdive.backend.security.authentication.domain.UserDetails;
+import com.devdive.backend.security.authentication.application.port.in.UserDetailsService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import org.apache.http.entity.ContentType;
